Sql_logs/requirements.txt
Flask
pyodbc
flask-cors
python-dotenv
elasticsearch

